Refactored embedded

See merge request polytechnique-montr-al/inf3995/20221/equipe-106/INF3995-106!116

+++ embedded/shared-firmware/src/vector3d.cc
@@ -48,7 +48,7 @@ Vector3D Vector3D::toUnitVector() const {
 }
 
 float Vector3D::length() const {
-  return static_cast<float>(std::sqrt(m_x * m_x + m_y * m_y + m_z * m_z));
+  return std::sqrt(m_x * m_x + m_y * m_y + m_z * m_z);
 }
 
 Vector3D& Vector3D::operator+=(const Vector3D& other) {
